Refactor frontend structure: move components to screens, add new header and styling, and implement user authentication flow. Remove old Home and Login components, and create new ones with improved layout and functionality.
