+++ frontend/src/App.css
@@ -1,3 +1,6 @@
 .App {
   width: 100%;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
 }
