+++ frontend/src/screens/Home.css
@@ -1,9 +1,15 @@
-.home-container {
-  max-width: 800px;
+.page-container {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
   width: 100%;
+  padding: 0;
+}
+
+.home-container {
+  width: 100%;
+  flex: 1;
   background: white;
-  border-radius: 16px;
-  box-shadow: 0 20px 60px rgba(0, 0, 0, 0.3);
   padding: 40px;
 }
 
@@ -115,4 +121,3 @@
   background: #e8e8e8;
   transform: translateY(-2px);
 }
-
